[refac] Refactor security package structure and update imports for better organization
