+++ edukit-api/src/main/java/com/edukit/api/security/authentication/MemberRoleConverter.java
@@ -1,4 +1,4 @@
-package com.edukit.api.converter;
+package com.edukit.api.security.authentication;
 
 import com.edukit.core.member.enums.MemberRole;
 import java.util.Collection;
@@ -7,6 +7,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 public class MemberRoleConverter {
+
     public static Collection<? extends GrantedAuthority> toGrantedAuthorities(final MemberRole memberRole) {
         return List.of(new SimpleGrantedAuthority(memberRole.getRole()));
     }
