+++ edukit-api/src/main/java/com/edukit/api/security/jwt/setting/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.edukit.api.security.jwt.config;
+package com.edukit.api.security.jwt.setting;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
